Refactor to move functions and its declarations to appropriate files
